# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pathfinder API",
    "description": "State-aware agentic trip planner. The frontend's only coupling is this contract (and the shared TripPlan JSON Schema in contracts/).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sessions": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Create Session",
        "operationId": "create_session_api_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "Get Session",
        "operationId": "get_session_api_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/context": {
      "put": {
        "tags": [
          "sessions"
        ],
        "summary": "Update Context",
        "operationId": "update_context_api_sessions__session_id__context_put",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Post Message",
        "operationId": "post_message_api_sessions__session_id__messages_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/plan/confirm": {
      "post": {
        "tags": [
          "plan"
        ],
        "summary": "Confirm Items",
        "operationId": "confirm_items_api_sessions__session_id__plan_confirm_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlan"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/sessions/{session_id}/plan/feedback": {
      "post": {
        "tags": [
          "plan"
        ],
        "summary": "Plan Feedback",
        "operationId": "plan_feedback_api_sessions__session_id__plan_feedback_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentName": {
        "type": "string",
        "enum": [
          "attraction",
          "hotel",
          "weather",
          "ticket"
        ],
        "title": "AgentName",
        "description": "The four pre-planning agents. Each owns one section of the TripPlan."
      },
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
...
      "Clarification": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "reason": {
            "$ref": "#/components/schemas/GateReason"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "reason"
        ],
        "title": "Clarification"
      },
      "ConfirmRequest": {
        "properties": {
          "item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Item Ids"
          },
          "ticket_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ticket Ids"
          },
          "hotel": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotel"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ConfirmRequest"
      },
...
      "ConversationTurn": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "route": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Route"
              },
              {
                "type": "null"
              }
            ]
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ConversationTurn"
      },
...
      "CreateSessionRequest": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/TripContext"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateSessionRequest"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "FeedbackRequest": {
        "properties": {
          "useful": {
            "type": "boolean",
            "title": "Useful"
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "useful",
          "rating"
        ],
        "title": "FeedbackRequest"
      },
      "FeedbackResponse": {
        "properties": {
          "queued": {
            "type": "boolean",
            "title": "Queued"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "queued"
        ],
        "title": "FeedbackResponse"
      },
      "GateReason": {
        "type": "string",
        "enum": [
          "accepted",
          "below_threshold",
          "model_flagged",
          "unclear_route",
          "missing_fields",
          "untyped_output",
          "no_plan_to_modify"
        ],
        "title": "GateReason"
      },
      "GateResult": {
        "properties": {
          "route": {
            "$ref": "#/components/schemas/Route"
          },
          "reason": {
            "$ref": "#/components/schemas/GateReason"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RouterDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "route",
          "reason"
        ],
        "title": "GateResult",
        "description": "Final routing outcome after the clarification gate. `route` is the path actually taken:\nanything not accepted becomes `Route.UNCLEAR` (clarify and return the turn to the user)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreferenceProfile": {
        "properties": {
          "hotel_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotel Style"
          },
          "liked_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Liked Categories"
          },
          "avoided_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Avoided Categories"
          },
          "budget_hint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "party_size_hint": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Size Hint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PreferenceProfile",
        "description": "What the user liked, kept as session memory and persisted by the background writer."
      },
      "QuickAnswer": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "from_plan": {
            "type": "boolean",
            "title": "From Plan",
            "description": "True when the current TripPlan already held the fact"
          },
          "related_to_plan": {
            "type": "boolean",
            "title": "Related To Plan"
          },
          "tool_called": {
            "type": "boolean",
            "title": "Tool Called"
          },
          "unavailable": {
            "type": "boolean",
            "title": "Unavailable",
            "description": "The one tool call failed",
            "default": false
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "from_plan",
          "related_to_plan",
          "tool_called"
        ],
        "title": "QuickAnswer"
      },
...
      "Route": {
        "type": "string",
        "enum": [
          "plan",
          "modify",
          "ask",
          "unclear"
        ],
        "title": "Route",
        "description": "The only outputs the System 1 router may produce (proposal: typed intent)."
      },
...
      "SessionView": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "context": {
            "$ref": "#/components/schemas/TripContext"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ConversationTurn"
            },
            "type": "array",
            "title": "History"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferenceProfile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "context",
          "plan",
          "history",
          "preferences"
        ],
        "title": "SessionView"
      },
...
      "TripContext": {
        "properties": {
          "destination": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin",
            "description": "Departure city for train/flight tickets"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 60.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Days"
          },
          "party_size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Size"
          },
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "hotel_style": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotel Style"
          },
          "hard_constraints": {
            "items": {
              "$ref": "#/components/schemas/HardConstraint"
            },
            "type": "array",
            "title": "Hard Constraints"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TripContext"
      },
      "TripPlan": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "plan_id": {
            "type": "string",
            "minLength": 1,
            "title": "Plan Id"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "default": 1
          },
          "destination": {
            "type": "string",
            "minLength": 1,
            "title": "Destination"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "party_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Party Size"
          },
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/DayPlan"
            },
            "type": "array",
            "minItems": 1,
            "title": "Days"
          },
          "places": {
            "items": {
              "$ref": "#/components/schemas/Place"
            },
            "type": "array",
            "title": "Places"
          },
          "hotel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HotelStay"
              },
              {
                "type": "null"
              }
            ]
          },
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/TicketOption"
            },
            "type": "array",
            "title": "Tickets"
          },
          "reservations": {
            "items": {
              "$ref": "#/components/schemas/Reservation"
            },
            "type": "array",
            "title": "Reservations"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionState"
            },
            "type": "array",
            "title": "Sections"
          },
          "cost": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CostBreakdown"
              },
              {
                "type": "null"
              }
            ]
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/ConstraintViolation"
            },
            "type": "array",
            "title": "Violations"
          },
          "disruptions": {
            "items": {
              "$ref": "#/components/schemas/Disruption"
            },
            "type": "array",
            "title": "Disruptions"
          },
          "saved_trip_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Saved Trip Refs"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "destination",
          "start_date",
          "end_date",
          "party_size",
          "days"
        ],
        "title": "TripPlan"
      },
      "TurnError": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/TurnErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "TurnError"
      },
...
      "TurnMetrics": {
        "properties": {
          "llm_calls": {
            "type": "integer",
            "title": "Llm Calls",
            "default": 0
          },
          "tool_calls": {
            "type": "integer",
            "title": "Tool Calls",
            "default": 0
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms",
            "default": 0.0
          },
          "agents_run": {
            "items": {
              "$ref": "#/components/schemas/AgentName"
            },
            "type": "array",
            "title": "Agents Run"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TurnMetrics"
      },
      "TurnResult": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "route": {
            "$ref": "#/components/schemas/Route"
          },
          "gate": {
            "$ref": "#/components/schemas/GateResult"
          },
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan_changed": {
            "type": "boolean",
            "title": "Plan Changed",
            "default": false
          },
          "answer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuickAnswer"
              },
              {
                "type": "null"
              }
            ]
          },
          "clarification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Clarification"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TurnError"
              },
              {
                "type": "null"
              }
            ]
          },
          "input_flagged": {
            "type": "boolean",
            "title": "Input Flagged",
            "default": false
          },
          "agents_run": {
            "items": {
              "$ref": "#/components/schemas/AgentName"
            },
            "type": "array",
            "title": "Agents Run"
          },
          "metrics": {
            "$ref": "#/components/schemas/TurnMetrics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "trace_id",
          "route",
          "gate",
          "reply"
        ],
        "title": "TurnResult"
      },
      "UpdateContextRequest": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/TripContext"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "context"
        ],
        "title": "UpdateContextRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}